Added Exceptions en ErrorsHandlers on Subscription

+++ src/service/subscription.service.ts
@@ -15,30 +15,41 @@ export class SubscriptionService{
         return  await this.subscriptionRepository.all();
     }
 
-    public async store(entry: SubscriptionCreateDto): Promise<void>{
+    public async store(entry: SubscriptionCreateDto): Promise<Subscription>{
         const originalEntry = await this.subscriptionRepository.findByUserAndCode(entry.user_id, entry.code);
 
         if(!originalEntry){
-            this.subscriptionRepository.store(entry as Subscription);
+            const subscriptionStored = await this.subscriptionRepository.store(entry as Subscription);            
+            return subscriptionStored;
         }else{
             throw new ApplicationException('user subscription already exists.');
-        }        
+        }                
     }
 
-    public async update(id:number, entry: SubscriptionUpdateDto): Promise<void>{
+    public async update(id:number, entry: SubscriptionUpdateDto): Promise<Subscription>{
         let originalEntry = await this.subscriptionRepository.find(id);
 
         if(originalEntry){
-           originalEntry.code = entry.code;
-           originalEntry.amount = entry.amount;
-           originalEntry.cron = entry.cron ;
+            //Se actualizan las propiedades
+          entry.amount ? originalEntry.amount = entry.amount : originalEntry.amount = originalEntry.amount;
+          entry.code ? originalEntry.code = entry.code : originalEntry.code = originalEntry.code;
+          entry.cron ? originalEntry.cron = entry.cron : originalEntry.cron = originalEntry.cron;
         }else{
             throw new ApplicationException('user subscription was not found');
-        }
+        }        
+        const updatedSubscription = await this.subscriptionRepository.update(originalEntry);
 
+        return updatedSubscription;
     }
 
     public async remove(id: number): Promise<void>{
-        return await this.subscriptionRepository.remove(id);
+        const originalEntry = await this.subscriptionRepository.find(id);
+
+        if(originalEntry){
+            return await this.subscriptionRepository.remove(id);                     
+        }else{
+            throw new ApplicationException('user subscription doesnt exists.');
+        }  
+        
     }
 };
